requirements.txt
# Azure dependencies
azure-search-documents==11.4.0
azure-core==1.30.0
openai==1.51.0

# PDF processing
PyPDF2==3.0.1

# Utilities
python-dotenv==1.0.0
